requirements.txt
crewai
streamlit
python-dotenv
google-generativeai
langchain-google-genai
